refactor: domain 을 직접 view에 넘기는 것이 아닌 dto를 통해 넘기도록 수정
